Fix levelup sound?
The problem is that *sometimes* the levelup sound doesn't happen.  Most
likely related to the frame counter.  The logic indicates that the
levelup sound effect happens after the tetris or line clear sound
effects.  What is most confusing is that the 30 wide big pieces hack
uses the same logic but doesn't have the same problem (or it does and I
haven't noticed?).

This seems to have fixed it by pushing the sound effect loading two
frames out.